student -pocket -bot/requirements.txt
flask
flask_sqlalchemy
flask_cors
pymysql
werkzeug
scikit-learn
numpy
reportlab
